# swagger_v1.yaml
openapi: 3.0.3
info:
  title: TravelGuide API
  description: API для управления городами и пользователями
  version: 2.0.0
servers:
  - url: http://localhost:8800/api/v2
    description: Development server
paths:
...
  /api/v1/users:
    post:
      summary: Создать администратора
      description: Создает нового администратора в системе
      operationId: register_admin
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                fio:
                  type: string
                  description: ФИО пользователя
                  example: "Иванов Иван Иванович"
                number_passport:
                  type: string
                  description: Номер паспорта
                  example: "1234 567890"
                phone_number:
                  type: string
                  description: Номер телефона
                  example: "+7 999 123-45-67"
                email:
                  type: string
                  format: email
                  description: Email адрес
                  example: "admin@example.com"
                login:
                  type: string
                  description: Логин для входа
                  example: "admin"
                password:
                  type: string
                  format: password
                  description: Пароль
                  example: "securepassword123"
              required:
                - fio
                - number_passport
                - phone_number
                - email
                - login
                - password
      responses:
        '201':
          description: Администратор успешно создан
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - users
    get:
      summary: Получить всех пользователей
      description: Возвращает список всех пользователей системы
      operationId: get_all_users
      responses:
        '201':
          description: Список пользователей получен успешно
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - users
  /api/v1/users/{user_id}:
    put:
      summary: Обновить администратора
      description: Обновляет информацию об администраторе
      operationId: update_admin
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                fio:
                  type: string
                  description: ФИО пользователя
                  example: "Иванов Иван Иванович"
                number_passport:
                  type: string
                  description: Номер паспорта
                  example: "1234 567890"
                phone_number:
                  type: string
                  description: Номер телефона
                  example: "+7 999 123-45-67"
                email:
                  type: string
                  format: email
                  description: Email адрес
                  example: "admin@example.com"
              required:
                - fio
                - number_passport
                - phone_number
                - email
      responses:
        '201':
          description: Администратор успешно обновлен
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - users

    delete:
      summary: Удалить пользователя
      description: Удаляет пользователя по ID
      operationId: delete_user
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - users
    get:
      summary: Получить профиль пользователя
      description: Возвращает детальную информацию о профиле пользователя с маршрутами
      operationId: get_user_profile
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID пользователя
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '201':
          description: Профиль пользователя получен успешно
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      tags:
        - users

  # ==================== АУТЕНТИФИКАЦИЯ ====================
...
  /api/v1/accommodations:
    get:
      summary: Get all accommodations
      description: Retrieve all accommodations and render HTML page
      operationId: get_all_accommodations
      tags:
        - accommodations
      responses:
        '201':
          description: HTML page with accommodations
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal server error

    post:
      summary: Create new accommodation
      description: Create a new accommodation and return HTML response
      operationId: create_accommodation
      tags:
        - accommodations
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                city_id:
                  type: integer
                check_in:
                  type: string
                  format: date-time
                check_out:
                  type: string
                  format: date-time
                address:
                  type: string
                cost:
                  type: number
                description:
                  type: string
      responses:
        '201':
          description: Accommodation created successfully
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad request
...
  /api/v1/entertainments:
    get:
      summary: Get all entertainments
      description: Retrieve all entertainments and render HTML page
      operationId: get_all_entertainments
      tags:
        - entertainments
      responses:
        '201':
          description: HTML page with entertainments
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal server error

    post:
      summary: Create new entertainment
      description: Create a new entertainment and return HTML response
      operationId: create_entertainment
      tags:
        - entertainments
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                city_id:
                  type: integer
                event_time:
                  type: string
                  format: date-time
                address:
                  type: string
                cost:
                  type: number
                description:
                  type: string
      responses:
        '201':
          description: Entertainment created successfully
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad request
...
  /api/v1/travels:
    get:
      summary: Get all travels
      description: Retrieve all travels with users, entertainments and accommodations
      operationId: get_all_travels
      tags:
        - travels
      responses:
        '201':
          description: HTML page with travels data
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal server error

    post:
      summary: Create new travel
      description: Create a new travel and return HTML response
      operationId: create_travel
      tags:
        - travels
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the travel
                description:
                  type: string
                  description: Description of the travel
                user_ids:
                  type: array
                  items:
                    type: integer
                  description: List of user IDs
                entertainment_ids:
                  type: array
                  items:
                    type: integer
                  description: List of entertainment IDs
                accommodation_ids:
                  type: array
                  items:
                    type: integer
                  description: List of accommodation IDs
      responses:
        '201':
          description: Travel created successfully
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad request
  /api/v1/travels/{travel_id}:
    put:
      summary: Update travel
      description: Update an existing travel and return HTML response
      operationId: update_travel
      tags:
        - travels
      parameters:
        - name: travel_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [planned, active, completed, cancelled]
                user_ids:
                  type: array
                  items:
                    type: integer
                entertainment_ids:
                  type: array
                  items:
                    type: integer
                accommodation_ids:
                  type: array
                  items:
                    type: integer
      responses:
        '201':
          description: Travel updated successfully
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Travel not found
        '500':
          description: Internal server error

    delete:
      summary: Delete travel
      description: Delete a travel and redirect to travels page
      operationId: delete_travel
      tags:
        - travels
      parameters:
        - name: travel_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '404':
          description: Travel not found
        '500':
          description: Internal server error

    patch:
      summary: Complete travel
      description: Mark a travel as completed and redirect to user profile
      operationId: complete_travel
      tags:
        - travels
      parameters:
        - name: travel_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '404':
          description: Travel or user not found
          content:
            text/html:
              schema:
                type: string
                example: "<h1>Пользователь не найден</h1>"
        '500':
          description: Internal server error

# ==================== МАРШРУТ ====================
...